avoid overridable method call in constructor for hover effect
