+++ src/io/bibleget/VersionsSelect.java
@@ -9,9 +9,6 @@ import ca.odell.glazedlists.BasicEventList;
 import ca.odell.glazedlists.EventList;
 import ca.odell.glazedlists.SeparatorList;
 import ca.odell.glazedlists.swing.DefaultEventListModel;
-import java.awt.Point;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
 import java.io.StringReader;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -24,7 +21,6 @@ import java.util.Set;
 import javax.json.Json;
 import javax.json.JsonObject;
 import javax.json.JsonReader;
-import javax.swing.DefaultListCellRenderer;
 import javax.swing.DefaultListSelectionModel;
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -40,7 +36,6 @@ public class VersionsSelect extends javax.swing.JList {
     private int versionLangs = 0;
     private int versionCount = 0;
     private VersionCellRenderer renderer;
-    private int mHoveredJListIndex = -1;
     
     private static DBHelper biblegetDB;
         
@@ -112,12 +107,11 @@ public class VersionsSelect extends javax.swing.JList {
             this.setModel(new DefaultEventListModel<>(versionsByLang));
             this.renderer = new VersionCellRenderer();
             this.setCellRenderer(renderer);
-            MouseAdapter mAdapter = new HoverMouseHandler(renderer);
+            
             
             this.setSelectionModel(new DisabledItemSelectionModel());
             if(loadPreferred){
                 this.setSelectedIndices(indices);
-                this.addMouseMotionListener(mAdapter);
             }
         }
     }
@@ -133,6 +127,10 @@ public class VersionsSelect extends javax.swing.JList {
     
     public int getVersionLangs(){
         return versionLangs;
+    }
+    
+    public VersionCellRenderer getRenderer(){
+        return renderer;
     }
     
     private static class VersionComparator implements Comparator<BibleVersion> {
@@ -189,36 +187,6 @@ public class VersionsSelect extends javax.swing.JList {
                 if(enabledFlags[index0]){ super.addSelectionInterval(index0,index0); }
             }
         }        
-        
-    }
-    
-    private class HoverMouseHandler extends MouseAdapter {
-        private final VersionCellRenderer renderer;
-        private int hoverIndex = -1;
-        
-        public HoverMouseHandler(VersionCellRenderer renderer){
-            this.renderer = renderer;
-        }
-        
-        @Override
-        public void mouseExited(MouseEvent e) {
-          setHoverIndex(-1);
-        }
-
-        @Override
-        public void mouseMoved(MouseEvent e) {
-          int index = VersionsSelect.this.locationToIndex(e.getPoint());
-          setHoverIndex(VersionsSelect.this.getCellBounds(index, index).contains(e.getPoint())
-                  ? index : -1);
-        }
-
-        private void setHoverIndex(int index) {
-          if (hoverIndex == index) return;
-          hoverIndex = index;
-          renderer.setHoverIndex(index);
-          VersionsSelect.this.repaint();
-        }
-        
         
     }
